Move `file-tree` to `file-explorer/views`
