+++ src/components/layout/main-sidebar.tsx
@@ -2,6 +2,7 @@ import { FilePlus, FolderOpen, FolderPlus, Server } from "lucide-react";
 import type React from "react";
 import { useMemo } from "react";
 import { cn } from "@/utils/cn";
+import FileTree from "../../file-explorer/views/file-tree";
 import { usePersistentSettingsStore } from "../../settings/stores/persistent-settings-store";
 import { useBufferStore } from "../../stores/buffer-store";
 import { useFileSystemStore } from "../../stores/file-system/store";
@@ -9,7 +10,6 @@ import { useProjectStore } from "../../stores/project-store";
 import { useSidebarStore } from "../../stores/sidebar-store";
 import { useUIState } from "../../stores/ui-state-store";
 import type { FileEntry } from "../../types/app";
-import FileTree from "../file-tree/file-tree";
 import GitView from "../git/git-view";
 import RemoteConnectionView from "../remote/remote-connection-view";
 import SearchView from "../search-view";
